Added count file function, quantile output and graph, readme and requirements update

+++ TFEA/main.py
@@ -14,6 +14,7 @@ def run():
     #Imports
     #==============================================================================
     import sys
+    import os
     import subprocess
     import shutil
     from pathlib import Path
@@ -149,7 +150,7 @@ def run():
         
     #ENRICHMENT module
     #==============================================================================
-    '''Where the bulk of TFEA analysis occurs. Some components of plotting module 
+    '''Where the bulk of TFEA and leading edgeanalysis occurs. Some components of plotting module 
         are contained within this enrichment module
     '''
     from TFEA import enrichment
@@ -164,15 +165,14 @@ def run():
 
     print("TFEA done. Output in:", config.vars['OUTPUT'], file=sys.stderr)
 
-    #LEADING EDGE module
-    #==============================================================================
-    '''A module to run the Leading edge on TFEA significance calls
-    '''
-
     #Delete temp_files directory
     #==============================================================================
     if not config.vars['DEBUG']:
+        os.remove("".join([config.vars['TEMPDIR'], "/ranked_file.fa"]))
+    if not config.vars['DEBUG'] and not config.vars['KEEP_LE_FILES']:
         shutil.rmtree(config.vars['TEMPDIR'])
+    
+    
 
     #==============================================================================
     #END OF MAIN SCRIPT
